move parser function to utils
