+++ pytorch_lightning/utilities/device_parser.py
@@ -16,6 +16,7 @@ from typing import Any, List, MutableSequence, Optional, Tuple, Union
 import torch
 
 from pytorch_lightning.plugins.environments import TorchElasticEnvironment
+from pytorch_lightning.tuner.auto_gpu_select import pick_multiple_gpus
 from pytorch_lightning.utilities import _TPU_AVAILABLE
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 
@@ -46,6 +47,20 @@ def determine_root_gpu_device(gpus: List[int]) -> Optional[int]:
     root_gpu = gpus[0]
 
     return root_gpu
+
+
+def _parse_devices(
+    gpus: Optional[Union[List[int], str, int]],
+    auto_select_gpus: bool,
+    tpu_cores: Optional[Union[List[int], str, int]],
+) -> Tuple[Optional[List[int]], Optional[Union[List[int], int]]]:
+    if auto_select_gpus and isinstance(gpus, int):
+        gpus = pick_multiple_gpus(gpus)
+
+    # TODO (@seannaren, @kaushikb11): Include IPU parsing logic here
+    gpu_ids = parse_gpu_ids(gpus)
+    tpu_cores = parse_tpu_cores(tpu_cores)
+    return gpu_ids, tpu_cores
 
 
 def parse_gpu_ids(gpus: Optional[Union[int, str, List[int]]]) -> Optional[List[int]]:
